Moving the various spec.aspect/blade/bladeset/workbench packages under a single package 'spec.model'
